% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_layers.R
\name{bound_mech_low}
\alias{bound_mech_low}
\title{Mechanical internal boundary layer; lowest height.}
\usage{
bound_mech_low(dist)
}
\arguments{
\item{dist}{Numeric. The distance to the point of roughness change in meters (m).}
}
\value{
Numeric. The height of the boundary layer in meters (m).
}
\description{
This function calculates the lowest height of the mechanical internal boundary layer (MIBL).
The MIBL height is calculated based on the distance to the point of roughness change using
an empirical relationship described by Bendix (2004, p. 242).
}
\details{
The mechanical internal boundary layer (MIBL) is a concept used in meteorology to describe the
layer of air that develops after a change in surface roughness. The height of the MIBL is
important for understanding the vertical distribution of wind speed and other meteorological
variables.

The function uses the formula:
\deqn{height = 0.3 \sqrt{dist}}
where \eqn{dist} is the distance to the point of roughness change in meters.
}
\examples{
# Calculate the lowest height of the MIBL for a distance of 100 meters
bound_mech_low(100)

}
\references{
Bendix, J. (2004). Weather and Climate: An Introduction. Springer.
}
