% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_station.R
\name{build_weather_station}
\alias{build_weather_station}
\title{Build a weather station object}
\usage{
build_weather_station(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link{rad_sw_toa.default}}, \code{\link{pres_p.default}}, \code{\link{trans_ozone.default}}, \code{\link{trans_aerosol.default}}, \code{\link{pres_sat_vapor_p.default}}, \code{\link{rad_lw_in.default}}
  \describe{
    \item{\code{sol_const}}{Solar radiation constant in W/m², default is 1361.}
    \item{\code{p0}}{Standard pressure in hPa, default 1013.25.}
    \item{\code{g}}{Gravitational acceleration in \eqn{m \cdot s^{-2}}, default 9.81.}
    \item{\code{rl}}{Specific gas constant for air in \eqn{m^2 \cdot s^{-2} \cdot K^{-1}}, default 287.05.}
    \item{\code{ozone_column}}{Atmospheric ozone as column in cm, default \code{ozone_column_default}.}
    \item{\code{vis}}{Visibility in km, default \code{vis_default}.}
    \item{\code{a}}{Constant a, default is 7.5 over water.}
    \item{\code{b}}{Constant b, default is 235 over water.}
  }}

\item{weather_station}{Object of class \code{weather_station}.}

\item{datetime}{Datetime of class \code{POSIXlt}. See \code{\link[base:as.POSIXlt]{base::as.POSIXlt()}}.
Make sure to provide the correct timezone information!}

\item{lon}{Longitude in degree.}

\item{lat}{Latitude in degree.}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Air temperature in degree Celcius.}

\item{t1}{Air temperature at lower height in degree Celcius.}

\item{t2}{Air temperature at upper height in degree Celcius.}

\item{v1}{Windspeed at lower height (e.g. height of anemometer) in m/s.}

\item{v2}{Windspeed at upper height in m/s.}

\item{slope}{Slope in degree.}

\item{exposition}{Exposition in degree.}

\item{surface_type}{Surface type.
Allowed values are: field, acre, lawn, street, agriculture, settlement, coniferous forest, deciduous forest, mixed forest, city, water, shrub.
EXCEPTION: for functions related to Priestley-Taylor methods,
allowed values are: field, bare soil, coniferous forest, water, wetland, spruce forest.}

\item{obs_height}{Height of obstacle in m.}

\item{valley}{Is the position in a valley (\code{TRUE}) or on a slope (\code{FALSE})?}

\item{surface_temp}{Surface temperature in degree Celcius.}

\item{rh}{Relative humidity in \%.}

\item{hum1}{Relative humidity at lower height in \%.}

\item{hum2}{Relative humidity at upper height in \%.}

\item{rad_bal}{Radiation balance in W/m\eqn{^2}.}

\item{texture}{Soil texture. Either "sand", "clay", or "peat".}

\item{moisture}{Soil moisture in cubic meter/cubic meter.}

\item{soil_flux}{Soil flux in W/m\eqn{^2}.}

\item{soil_temp1}{Soil temperature in degree Celcius of measurement 1.}

\item{soil_temp2}{Soil temperature in degree Celcius of measurement 2.}

\item{soil_depth1}{Depth of the soil temperature measurement 1 in m.}

\item{soil_depth2}{Depth of the soil temperature measurement 2 in m.}
}
\description{
Creates a list of class \code{weather_station} that contains all input arguments.
}
\details{
Provided input arguments will only be used if they are listed in the section
"Arguments". No warning message is generated for unused arguments.
}
