% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldClim-package.R
\docType{package}
\name{fieldClim-package}
\alias{fieldClim}
\alias{fieldClim-package}
\title{fieldClim: Calculates Fluxes for Weather Stations}
\description{
`fieldClim` is designed as a handy tool for calculations of various weather and micromicrometeorological parameters based on the measurements of a weather station. It includes functions adressing the estimation of radiation properties, the calculations of latent, sensible, and turbulent heat fluxes, the calculation of soil heat fluxes as well as the estimation of thermal and mechanical boundary layers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gisma.github.io/migration-fieldclim}
  \item \url{https://github.com/gisma/migration-fieldclim}
  \item Report bugs at \url{https://github.com/gisma/migration-fieldclim/issues}
}

}
\author{
\strong{Maintainer}: Christoph Reudenbach \email{reudenbach@uni-marburg.de}

Authors:
\itemize{
  \item Christoph Reudenbach \email{reudenbach@uni-marburg.de}
  \item David Langenohl \email{langenoh@students.uni-marburg.de}
  \item Jonas Viehweger \email{jonas.viehweger@rwth-aachen.de}
  \item Jörg Bendix \email{bendix@geo.uni-marburg.de}
  \item Fabian Mitze \email{Mitze@students.uni-marburg.de}
  \item Hanna Wenzel \email{wenzel@students.uni-marburg.de}
  \item Kai Richter \email{richter5@students.uni-marburg.de}
  \item You-De Chen \email{you-de.chen@students.uni-marburg.de}
}

}
\keyword{internal}
