% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{hum_absolute}
\alias{hum_absolute}
\alias{hum_absolute.default}
\alias{hum_absolute.weather_station}
\title{Absolute humidity}
\usage{
hum_absolute(...)

\method{hum_absolute}{default}(rh, temp, ...)

\method{hum_absolute}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{rh}{Relative humidity in \%.}

\item{temp}{Temperature in °C.}

\item{weather_station}{A weather_station object.}
}
\value{
Numeric. Absolute humidity in kg/m³.
}
\description{
Calculates absolute humidity from vapor pressure and air temperature.
}
\details{
Absolute humidity (\eqn{AH}) is the mass of water vapor per unit volume of air. It is calculated from the vapor pressure and temperature using the formula:
\deqn{AH = \frac{0.21668 \times pvapor}{T}}
where \eqn{pvapor} is the vapor pressure and \eqn{T} is the temperature in Kelvin.
The formula and constants follow the Bendix (2004) humidity helper
background used by the package.
}
\examples{
# Calculate absolute humidity
hum_absolute(rh = 70, temp = 25)
}
\references{
Bendix 2004, p. 262.
}
