% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{hum_moisture_gradient}
\alias{hum_moisture_gradient}
\alias{hum_moisture_gradient.default}
\alias{hum_moisture_gradient.weather_station}
\title{Moisture gradient}
\usage{
hum_moisture_gradient(...)

\method{hum_moisture_gradient}{default}(hum1, hum2, t1, t2, z1 = 2, z2 = 10, elev, ...)

\method{hum_moisture_gradient}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{hum1}{Relative humidity at lower height in \%.}

\item{hum2}{Relative humidity at upper height in \%.}

\item{t1}{Air temperature at lower height in °C.}

\item{t2}{Air temperature at upper height in °C.}

\item{z1}{Lower measurement height in m.}

\item{z2}{Upper measurement height in m.}

\item{elev}{Elevation above sea level in m.}

\item{weather_station}{A weather_station object.}
}
\value{
Numeric. Moisture gradient.
}
\description{
Calculates moisture gradient.
}
\details{
The moisture gradient is calculated as the difference in specific humidity at two heights divided by the difference in heights:
\deqn{\Delta q / \Delta z}
where \eqn{\Delta q} is the difference in specific humidity and \eqn{\Delta z} is the difference in heights.
}
\examples{
# Calculate moisture gradient
hum_moisture_gradient(hum1 = 80, hum2 = 60, t1 = 20, t2 = 15, z1 = 2, z2 = 10, elev = 100)
}
