% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{pos_anemometer_height}
\alias{pos_anemometer_height}
\title{Necessary Anemometer Height}
\usage{
pos_anemometer_height(dist, min_dist, obs_height)
}
\arguments{
\item{dist}{Distance between the climate station and the obstacle (e.g., forest) in meters (m).}

\item{min_dist}{Minimum distance between the climate station and the obstacle required to ensure independent measurements, in meters (m).}

\item{obs_height}{Height of the obstacle in meters (m).}
}
\value{
A message indicating if the climate station is well positioned or, if not, the height at which the anemometer needs to be positioned.
}
\description{
Checks if the distance between the climate station and the forest is smaller than the minimum distance. If so, calculates the height at which the anemometer needs to be positioned to ensure independent measurements.
}
\details{
If the climate station is positioned closer than the minimum required distance, the anemometer needs to be positioned at a higher altitude to avoid measurement interference.

Formula used:
\eqn{h_{new} = h \cdot \frac{d_{min} - d}{d_{min}}}
}
\examples{
# Check anemometer height for a station too close to the obstacle
pos_anemometer_height(dist = 50, min_dist = 100, obs_height = 30)
}
