% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressure.R
\name{pres_p}
\alias{pres_p}
\alias{pres_p.default}
\alias{pres_p.weather_station}
\title{Air Pressure}
\usage{
pres_p(...)

\method{pres_p}{default}(elev, temp, ..., p0 = p0_default, g = g_default, rl = rl_default)

\method{pres_p}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Named station fields, site parameters or model assumptions.}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Air temperature in degrees C.}

\item{p0}{Standard pressure in hPa, default 1013.25.}

\item{g}{Gravitational acceleration in \eqn{m \cdot s^{-2}}, default 9.81.}

\item{rl}{Specific gas constant for air in \eqn{m^2 \cdot s^{-2} \cdot K^{-1}}, default 287.05.}

\item{weather_station}{A weather_station object.}
}
\value{
Air pressure in hPa.
}
\description{
Calculate air pressure based on the barometric formula.
}
\details{
The formula assumes that the temperature does not change with altitude.
The results are accurate for elevations lower than 5 km.

The air pressure (\eqn{p}) is calculated using the barometric formula:
\deqn{p = p_0 \cdot \exp{\left(-\frac{g \cdot h}{R \cdot T}\right)}}
where:
\eqn{p_0} is the standard pressure (default 1013.25 hPa),
\eqn{g} is the gravitational acceleration (default 9.80665 m/s²),
\eqn{h} is the elevation above sea level in meters (m),
\eqn{R} is the specific gas constant for air (default 287.05 m²/s²/K), and
\eqn{T} is the temperature in Kelvin (K).
}
\examples{
# Calculate air pressure at an elevation of 500 meters and temperature of 15°C
pres_p(elev = 500, temp = 15)
}
\references{
Lente & Ősz 2020 eq. 5.
}
