% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressure.R
\name{pres_sat_vapor_p}
\alias{pres_sat_vapor_p}
\alias{pres_sat_vapor_p.default}
\alias{pres_sat_vapor_p.weather_station}
\title{Saturated Vapor Pressure}
\usage{
pres_sat_vapor_p(...)

\method{pres_sat_vapor_p}{default}(temp, ..., a = 7.5, b = 235)

\method{pres_sat_vapor_p}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Named station fields, site parameters or model assumptions.}

\item{temp}{Air temperature in degrees C.}

\item{a}{Constant a, default is 7.5 over water.}

\item{b}{Constant b, default is 235 over water.}

\item{weather_station}{A weather_station object.}
}
\value{
Saturation vapor pressure in hPa.
}
\description{
Calculates the saturation vapor pressure for a given temperature.
}
\details{
The saturation vapor pressure (\eqn{e_s}) is calculated using the formula:
\deqn{e_s = 6.1078 \cdot 10^{\left(\frac{a \cdot T}{b + T}\right)}}
where:
\eqn{T} is the temperature in °C,
\eqn{a} and \eqn{b} are constants that vary depending on the state of water:
\itemize{
\item Over water: \eqn{a = 7.5}, \eqn{b = 235}
\item Over undercooled water: \eqn{a = 7.6}, \eqn{b = 240.7}
\item Over ice: \eqn{a = 9.5}, \eqn{b = 265.5}
}
}
\examples{
# Calculate saturation vapor pressure at a temperature of 20°C
pres_sat_vapor_p(temp = 20)
}
\references{
Bendix 2004, p. 261.
}
