% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressure.R
\name{pres_vapor_p}
\alias{pres_vapor_p}
\alias{pres_vapor_p.default}
\alias{pres_vapor_p.weather_station}
\title{Vapor Pressure}
\usage{
pres_vapor_p(...)

\method{pres_vapor_p}{default}(temp, rh, ...)

\method{pres_vapor_p}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{temp}{Air temperature in degree Celcius.}

\item{rh}{Relative humidity in \%.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Vapor pressure in hPa.
}
\description{
Calculates vapor pressure from relative humidity and saturation vapor pressure.
}
\details{
The vapor pressure (\eqn{e}) is calculated as:
\deqn{e = \frac{RH}{100} \cdot e_s}
where:
\eqn{RH} is the relative humidity in \%,
\eqn{e_s} is the saturation vapor pressure in hPa.
}
\examples{
# Calculate vapor pressure at a temperature of 20°C and 60\% relative humidity
pres_vapor_p(temp = 20, rh = 60)
}
\references{
Bendix 2004, p. 262
}
