% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_bal}
\alias{rad_bal}
\alias{rad_bal.default}
\alias{rad_bal.weather_station}
\title{Total Radiation Balance}
\usage{
rad_bal(...)

\method{rad_bal}{default}(
  datetime,
  lon,
  lat,
  elev,
  temp,
  rh,
  slope,
  exposition,
  valley,
  surface_type,
  surface_temp,
  ...
)

\method{rad_bal}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime of class \code{POSIXlt}. See \code{\link[base:as.POSIXlt]{base::as.POSIXlt()}}.
Make sure to provide the correct timezone information!}

\item{lon}{Longitude in degree.}

\item{lat}{Latitude in degree.}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Air temperature in degree Celcius.}

\item{rh}{Relative humidity in \%.}

\item{slope}{Slope in degree.}

\item{exposition}{Exposition in degree.}

\item{valley}{Is the position in a valley (\code{TRUE}) or on a slope (\code{FALSE})?}

\item{surface_type}{Surface type.
Allowed values are: field, acre, lawn, street, agriculture, settlement, coniferous forest, deciduous forest, mixed forest, city, water, shrub.
EXCEPTION: for functions related to Priestley-Taylor methods,
allowed values are: field, bare soil, coniferous forest, water, wetland, spruce forest.}

\item{surface_temp}{Surface temperature in degree Celcius.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Total radiation balance in W/m².
}
\description{
Calculates the total radiation balance by summing the shortwave and longwave radiation balances.
}
\details{
The total radiation balance (\eqn{R_{total}}) is calculated as:
\deqn{R_{total} = R_{sw} + R_{lw}}
where:
\eqn{R_{sw}} is the shortwave radiation balance,
\eqn{R_{lw}} is the longwave radiation balance.
}
\examples{
# Calculate total radiation balance
rad_bal(datetime = Sys.time(), lon = 10, lat = 50, elev = 100, temp = 15, rh = 60,
        slope = 5, exposition = 180, valley = FALSE, surface_type = "lawn", surface_temp = 15)
}
\references{
Bendix 2004, p. 45 eq. 3.1.
}
