% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_emissivity_air}
\alias{rad_emissivity_air}
\alias{rad_emissivity_air.default}
\alias{rad_emissivity_air.weather_station}
\title{Emissivity of the Atmosphere}
\usage{
rad_emissivity_air(...)

\method{rad_emissivity_air}{default}(temp, rh, ...)

\method{rad_emissivity_air}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{temp}{Air temperature in degree Celcius.}

\item{rh}{Relative humidity in \%.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Emissivity ratio from 0 to 1.
}
\description{
Calculates the emissivity of the atmosphere.
}
\details{
The emissivity of the atmosphere (\eqn{\epsilon_{air}}) is calculated as:
\deqn{\epsilon_{air} = (1.24 \cdot \frac{e}{T_{air}})^{1/7}}
where:
\eqn{e} is the vapor pressure,
\eqn{T_{air}} is the air temperature in Kelvin.
}
\examples{
# Calculate emissivity of the atmosphere
rad_emissivity_air(temp = 15, rh = 60)
}
\references{
Bendix 2004, p. 66 eq. 3.22.
}
