% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_lw_in}
\alias{rad_lw_in}
\alias{rad_lw_in.default}
\alias{rad_lw_in.weather_station}
\title{Longwave Incoming Radiation}
\usage{
rad_lw_in(...)

\method{rad_lw_in}{default}(temp, rh, slope, valley, ..., sigma = sigma_default)

\method{rad_lw_in}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Named station fields, site parameters or model assumptions.}

\item{temp}{Air temperature in degrees C.}

\item{rh}{Relative humidity in percent.}

\item{slope}{Slope in degrees.}

\item{valley}{Logical value indicating whether the station is in a valley.}

\item{sigma}{Stefan-Boltzmann constant in W m-2 K-4.}

\item{weather_station}{A weather_station object.}
}
\value{
Longwave incoming radiation in W/m².
}
\description{
Calculates the longwave radiation of the atmosphere.
}
\details{
The longwave incoming radiation (\eqn{LW_{in}}) is calculated as:
\deqn{LW_{in} = \epsilon_{air} \cdot \sigma \cdot T_{air}^4 \cdot \text{sky\_view}}
where:
\eqn{\epsilon_{air}} is the emissivity of the air,
\eqn{\sigma} is the Stefan-Boltzmann constant,
\eqn{T_{air}} is the air temperature in Kelvin, and
\eqn{\text{sky\_view}} is the sky view factor.
}
\examples{
# Calculate longwave incoming radiation
rad_lw_in(temp = 15, rh = 60, slope = 5, valley = FALSE)
}
\references{
Bendix 2004, p. 68 eq. 3.24.
}
