% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\name{soil_heat_flux}
\alias{soil_heat_flux}
\alias{soil_heat_flux.default}
\alias{soil_heat_flux.weather_station}
\title{Soil Heat Flux}
\usage{
soil_heat_flux(...)

\method{soil_heat_flux}{default}(
  texture,
  moisture,
  soil_temp1,
  soil_temp2,
  soil_depth1,
  soil_depth2,
  ...
)

\method{soil_heat_flux}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{texture}{Soil texture. Either "sand", "peat" or "clay".}

\item{moisture}{Soil moisture content in cubic meters per cubic meter.}

\item{soil_temp1}{Temperature at the first depth in °C.}

\item{soil_temp2}{Temperature at the second depth in °C.}

\item{soil_depth1}{Depth of the first measurement in m.}

\item{soil_depth2}{Depth of the second measurement in m.}

\item{weather_station}{A weather_station object.}
}
\value{
Soil heat flux in W/m².
}
\description{
Calculates soil heat flux from measurements in two different depths and thermal conductivity of the soil.
Negative values signify flux towards the atmosphere, while positive values signify flux into the soil.
}
\details{
The soil heat flux (\eqn{G}) is calculated using the formula:
\deqn{G = -\lambda \cdot \frac{T_1 - T_2}{z_1 - z_2}}
where:
\eqn{\lambda} is the thermal conductivity of the soil (W/m/K),
\eqn{T_1} and \eqn{T_2} are the temperatures at two different depths (°C),
\eqn{z_1} and \eqn{z_2} are the depths at which the temperatures are measured (m).
}
\examples{
# Calculate soil heat flux
soil_heat_flux(texture = "sand", moisture = 0.25, soil_temp1 = 15, soil_temp2 = 10, soil_depth1 = 0.1, soil_depth2 = 0.3)
}
\references{
Bendix 2004, p. 71 eq. 4.2.
}
