% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_azimuth}
\alias{sol_azimuth}
\alias{sol_azimuth.default}
\alias{sol_azimuth.weather_station}
\title{Solar azimuth}
\usage{
sol_azimuth(...)

\method{sol_azimuth}{default}(datetime, lon, lat, ...)

\method{sol_azimuth}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Named station fields, site parameters or model assumptions.}

\item{datetime}{POSIXlt or POSIXct date-time vector.}

\item{lon}{Longitude in degrees.}

\item{lat}{Latitude in degrees.}

\item{weather_station}{A weather_station object.}
}
\value{
Degree.
}
\description{
Calculates the solar azimuth, which is the compass direction from which the sunlight is coming at any specific point on the earth's surface.
}
\details{
The solar azimuth (\eqn{A}) is calculated as:
\deqn{A = \arccos\left(\frac{\sin(\delta) \cdot \cos(\phi) - \cos(\delta) \cdot \sin(\phi) \cdot \cos(H)}{\cos(h)}\right)}
where:
\eqn{\delta} is the solar declination,
\eqn{\phi} is the latitude,
\eqn{H} is the hour angle,
\eqn{h} is the solar elevation angle.
}
\examples{
# Calculate solar azimuth
sol_azimuth(as.POSIXlt("2022-06-21 12:00:00"), lon = 10, lat = 50)
}
\references{
Bendix 2004, p. 243.
}
