% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_day_angle}
\alias{sol_day_angle}
\alias{sol_day_angle.default}
\alias{sol_day_angle.weather_station}
\title{Day angle}
\usage{
sol_day_angle(...)

\method{sol_day_angle}{default}(datetime, ...)

\method{sol_day_angle}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime of class \code{POSIXlt}. See \code{\link[base:as.POSIXlt]{base::as.POSIXlt()}}.
Make sure to provide the correct timezone information!}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Degree.
}
\description{
See a year as a circle. The first day of a (leap) year is 0 degree
and the last day 360 degrees.
}
\details{
The day angle (\eqn{D}) is calculated as:
\deqn{D = \frac{2\pi(J - 1)}{365}}
where:
\eqn{J} is the Julian day.
}
\examples{
# Calculate day angle
sol_day_angle(as.POSIXlt("2022-06-21"))
}
\references{
Bendix 2004, p. 243.
}
