% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_declination}
\alias{sol_declination}
\alias{sol_declination.default}
\alias{sol_declination.weather_station}
\title{Solar declination}
\usage{
sol_declination(...)

\method{sol_declination}{default}(datetime, ...)

\method{sol_declination}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime of class \code{POSIXlt}. See \code{\link[base:as.POSIXlt]{base::as.POSIXlt()}}.
Make sure to provide the correct timezone information!}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Degree.
}
\description{
Solar declination
}
\details{
The solar declination (\eqn{\delta}) is the angle between the rays of the sun and the plane of the Earth's equator. It is calculated as:
\deqn{\delta = \arcsin(\sin(23.44^\circ) \cdot \sin(L))}
where:
\eqn{L} is the ecliptic longitude.
}
\examples{
# Calculate solar declination
sol_declination(as.POSIXlt("2022-06-21"))
}
\references{
Bendix 2004, p. 243.
}
