% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_ecliptic_length}
\alias{sol_ecliptic_length}
\alias{sol_ecliptic_length.default}
\alias{sol_ecliptic_length.weather_station}
\title{Solar ecliptic length}
\usage{
sol_ecliptic_length(...)

\method{sol_ecliptic_length}{default}(datetime, ...)

\method{sol_ecliptic_length}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime of class \code{POSIXlt}. See \code{\link[base:as.POSIXlt]{base::as.POSIXlt()}}.
Make sure to provide the correct timezone information!}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Degree.
}
\description{
Calculates the solar ecliptic length, which is the angle of the Earth's orbit around the sun relative to the vernal equinox.
}
\details{
The solar ecliptic length (\eqn{L}) is calculated as:
\deqn{L = 279.3 + 0.9856 \cdot J + 1.92 \cdot \sin(M)}
where:
\eqn{J} is the Julian day,
\eqn{M} is the solar medium anomaly in radians.
}
\examples{
# Calculate solar ecliptic length
sol_ecliptic_length(as.POSIXlt("2022-06-21"))
}
\references{
Bendix 2004, p. 243.
}
