% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_julian_day}
\alias{sol_julian_day}
\title{Julian day}
\usage{
sol_julian_day(...)
}
\arguments{
\item{...}{Named station fields, site parameters or model assumptions.}
}
\value{
Unitless.
}
\description{
Day of year as an integer from 1 to 366.
}
\details{
The Julian day (\eqn{J}) is the day of the year, ranging from 1 to 366.
}
\examples{
# Calculate Julian day
sol_julian_day(as.POSIXlt("2022-06-21"))
}
