% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_medium_suntime}
\alias{sol_medium_suntime}
\alias{sol_medium_suntime.default}
\alias{sol_medium_suntime.weather_station}
\title{Solar medium suntime}
\usage{
sol_medium_suntime(...)

\method{sol_medium_suntime}{default}(datetime, lon, ...)

\method{sol_medium_suntime}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime of class \code{POSIXlt}. See \code{\link[base:as.POSIXlt]{base::as.POSIXlt()}}.
Make sure to provide the correct timezone information!}

\item{lon}{Longitude in degree.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Hour.
}
\description{
Calculates the solar medium suntime, which is the mean solar time adjusted for the observer's longitude.
}
\details{
The solar medium suntime (\eqn{T_m}) is calculated as:
\deqn{T_m = T_{local} + \frac{lon}{15}}
where:
\eqn{T_{local}} is the local time zone,
\eqn{lon} is the longitude of the observer.
}
\examples{
# Calculate solar medium suntime
sol_medium_suntime(as.POSIXlt("2022-06-21 12:00:00"), lon = 10)
}
\references{
Bendix 2004, p. 243.
}
