% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_time_formula}
\alias{sol_time_formula}
\alias{sol_time_formula.default}
\alias{sol_time_formula.weather_station}
\title{Solar time formula}
\usage{
sol_time_formula(...)

\method{sol_time_formula}{default}(datetime, lon, ...)

\method{sol_time_formula}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime of class \code{POSIXlt}. See \code{\link[base:as.POSIXlt]{base::as.POSIXlt()}}.
Make sure to provide the correct timezone information!}

\item{lon}{Longitude in degree.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Hour.
}
\description{
Calculates the solar time formula, which corrects the solar medium suntime to account for the Earth's elliptical orbit and axial tilt.
}
\details{
The solar time formula (\eqn{E_t}) is calculated as:
\deqn{E_t = 0.1644 \cdot \sin(2L) - 0.1277 \cdot \sin(M)}
where:
\eqn{L} is the ecliptic longitude,
\eqn{M} is the solar medium anomaly.
}
\examples{
# Calculate solar time formula
sol_time_formula(as.POSIXlt("2022-06-21 12:00:00"), lon = 10)
}
\references{
Bendix 2004, p. 243.
}
