% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{terr_sky_view}
\alias{terr_sky_view}
\alias{terr_sky_view.default}
\alias{terr_sky_view.weather_station}
\title{Sky View Factor}
\usage{
terr_sky_view(...)

\method{terr_sky_view}{default}(slope, valley, ...)

\method{terr_sky_view}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments passed to other methods.}

\item{slope}{Slope of the terrain in degrees.}

\item{valley}{Logical value indicating if the location is in a valley (TRUE) or not (FALSE).}

\item{weather_station}{A weather_station object.}
}
\value{
Ratio from 0 to 1, unitless, representing the sky view factor.
}
\description{
Calculates the sky view factor, which represents how much sky can be seen from a given point.
}
\details{
The sky view factor (\eqn{SVF}) is calculated as:
\deqn{SVF = \frac{1 + \cos(\theta)}{2}} for non-valley locations,
and \deqn{SVF = \cos(\theta)} for valley locations,
where \eqn{\theta} is the slope angle.

The terrain view factor can be calculated by \code{1 - terr_sky_view()}, which represents how much terrain can be seen from the point.
}
\examples{
# Sky view factor for a slope of 30 degrees not in a valley
terr_sky_view(slope = 30, valley = FALSE)

# Sky view factor for a slope of 30 degrees in a valley
terr_sky_view(slope = 30, valley = TRUE)
}
\references{
Bendix 2004, p. 63 eq. 3.15.
}
