% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_gas}
\alias{trans_gas}
\alias{trans_gas.default}
\alias{trans_gas.weather_station}
\title{Transmittance due to gas}
\usage{
trans_gas(...)

\method{trans_gas}{default}(datetime, lon, lat, elev, temp, ...)

\method{trans_gas}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{POSIXlt object, date and time of the observation.}

\item{lon}{Longitude in decimal degrees.}

\item{lat}{Latitude in decimal degrees.}

\item{elev}{Elevation above sea level in meters.}

\item{temp}{Air temperature in °C.}

\item{weather_station}{A weather_station object.}
}
\value{
Transmittance ratio from 0 to 1, unitless.
}
\description{
Calculates transmittance due to O\eqn{_2} and CO\eqn{_2}.
}
\details{
The transmittance due to gases is calculated using the formula:
\deqn{T_{gas} = \exp(-0.0127 \cdot M_{abs}^{0.26})}
where \eqn{M_{abs}} is the absolute optical air mass.
This is an atmospheric transmittance approximation cited to Bendix (2004)
appendix material.
}
\examples{
# Calculate transmittance due to gas
trans_gas(
  datetime = as.POSIXlt("2023-08-06 12:00:00", tz = "UTC"),
  lon = 8.68, lat = 50.77, elev = 100, temp = 20
)
}
\references{
Bendix 2004, p. 246.
}
