% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_ozone}
\alias{trans_ozone}
\alias{trans_ozone.default}
\alias{trans_ozone.weather_station}
\title{Transmittance due to ozone}
\usage{
trans_ozone(...)

\method{trans_ozone}{default}(datetime, lon, lat, ..., ozone_column = ozone_column_default)

\method{trans_ozone}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{POSIXlt object, date and time of the observation.}

\item{lon}{Longitude in decimal degrees.}

\item{lat}{Latitude in decimal degrees.}

\item{ozone_column}{Atmospheric ozone as column in cm, default \code{ozone_column_default}.}

\item{weather_station}{A weather_station object.}
}
\value{
Transmittance ratio from 0 to 1, unitless.
}
\description{
Calculates transmittance due to atmospheric ozone.
}
\details{
The transmittance due to ozone is calculated using the formula:
\deqn{T_{ozone} = 1 - (0.1611 \cdot x \cdot (1 + 139.48 \cdot x)^{-0.3035} - 0.002715 \cdot x \cdot (1 + 0.044 \cdot x + 0.0003 \cdot x^2)^{-1})}
where \eqn{x} is the product of the ozone column and the relative optical air mass.
This is an Iqbal-style atmospheric transmittance approximation as presented
in the Bendix (2004) appendix material cited for the package.
}
\examples{
# Calculate transmittance due to ozone
trans_ozone(
  datetime = as.POSIXlt("2023-08-06 12:00:00", tz = "UTC"),
  lon = 8.68, lat = 50.77, ozone_column = 0.3
)
}
\references{
Bendix 2004, p. 245.
}
