% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_vapor}
\alias{trans_vapor}
\alias{trans_vapor.default}
\alias{trans_vapor.weather_station}
\title{Transmittance due to water vapor}
\usage{
trans_vapor(...)

\method{trans_vapor}{default}(datetime, lon, lat, elev, temp, ...)

\method{trans_vapor}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{POSIXlt object, date and time of the observation.}

\item{lon}{Longitude in decimal degrees.}

\item{lat}{Latitude in decimal degrees.}

\item{elev}{Elevation above sea level in meters.}

\item{temp}{Air temperature in °C.}

\item{weather_station}{A weather_station object.}
}
\value{
Transmittance ratio from 0 to 1, unitless.
}
\description{
Calculates transmittance due to water vapor.
}
\details{
The transmittance due to water vapor is calculated using the formula:
\deqn{T_{vapor} = 1 - 2.4959 \cdot x \cdot ((1 + 79.034 \cdot x)^{0.6828} + 6.385 \cdot x)^{-1}}
where \eqn{x} is the product of the precipitable water and the relative optical air mass.
This is an Iqbal-style atmospheric transmittance approximation as presented
in the Bendix (2004) appendix material cited for the package.
}
\examples{
# Calculate transmittance due to water vapor
trans_vapor(
  datetime = as.POSIXlt("2023-08-06 12:00:00", tz = "UTC"),
  lon = 8.68, lat = 50.77, elev = 100, temp = 20
)
}
\references{
Bendix 2004, p. 245.
}
