% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk.R
\name{turb_flux_bulk_residual}
\alias{turb_flux_bulk_residual}
\title{Bulk-residual turbulent heat flux workflow}
\usage{
turb_flux_bulk_residual(weather_station, ...)
}
\arguments{
\item{weather_station}{A \code{weather_station} object.}

\item{...}{Further arguments passed to \code{sensible_bulk()}.}
}
\value{
The input \code{weather_station} object with additional fields:
\code{sensible_bulk} and \code{latent_bulk_residual}.
}
\description{
Adds \code{sensible_bulk} and \code{latent_bulk_residual} to a \code{weather_station}
object.
}
