% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbulent_flux.R
\name{turb_flux_ex_quotient_imp}
\alias{turb_flux_ex_quotient_imp}
\alias{turb_flux_ex_quotient_imp.default}
\alias{turb_flux_ex_quotient_imp.weather_station}
\title{Exchange quotient for impulse transmission}
\usage{
turb_flux_ex_quotient_imp(...)

\method{turb_flux_ex_quotient_imp}{default}(
  t1,
  t2,
  z1 = 2,
  z2 = 10,
  v1,
  v2,
  elev,
  surface_type = NULL,
  obs_height = NULL,
  ...
)

\method{turb_flux_ex_quotient_imp}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{t1}{Temperature at lower height (e.g. height of anemometer) in °C.}

\item{t2}{Temperature at upper height in degrees C.}

\item{z1}{Lower height of measurement (e.g. height of anemometer) in m.}

\item{z2}{Upper height of measurement in m.}

\item{v1}{Windspeed at lower height (e.g. height of anemometer) in m/s.}

\item{v2}{Windspeed at upper height in m/s.}

\item{elev}{Elevation above sea level in m.}

\item{surface_type}{Type of surface. Options: r surface_properties$surface_type}

\item{obs_height}{Height of obstacle in meters (m).}

\item{weather_station}{Object of class weather_station}
}
\value{
Exchange quotient for impulse transmission in kg/(m*s).
}
\description{
Calculation of the exchange quotient of the turbulent impulse transmission.
The Businger-type constants use Foken (2016) method background and are
documented as the current package implementation.
}
\references{
Foken 2016, p. 361: Businger.
}
