% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbulent_flux.R
\name{turb_flux_monin}
\alias{turb_flux_monin}
\alias{turb_flux_monin.default}
\alias{turb_flux_monin.weather_station}
\title{Monin-Obhukov-Length}
\usage{
turb_flux_monin(...)

\method{turb_flux_monin}{default}(
  z1 = 2,
  z2 = 10,
  v1,
  v2,
  t1,
  t2,
  elev,
  surface_type = NULL,
  obs_height = NULL,
  ...
)

\method{turb_flux_monin}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{z1}{Lower height of measurement (e.g. height of anemometer) in m.}

\item{z2}{Upper height of measurement in m.}

\item{v1}{Windspeed at lower height (e.g. height of anemometer) in m/s.}

\item{v2}{Windspeed at upper height in m/s.}

\item{t1}{Temperature at lower height (e.g. height of anemometer) in °C.}

\item{t2}{Temperature at upper height in °C.}

\item{elev}{Elevation above sea level in m.}

\item{surface_type}{Type of surface. Options: r surface_properties$surface_type}

\item{obs_height}{Height of obstacle in meters (m).}

\item{weather_station}{Object of class weather_station}
}
\value{
Monin-Obhukov-Length in m.
}
\description{
Calculation of the Monin-Obhukov-Length.
The calculation depends on the stability of the atmosphere.
This value will be taken from the Gradient-Richardson-Number.
The profile and stability parameterization uses Bendix/Foken method
background.
}
\references{
Bendix 2004, p. 241
}
