% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{hum_evap_heat}
\alias{hum_evap_heat}
\alias{hum_evap_heat.default}
\alias{hum_evap_heat.weather_station}
\title{Enthalpy of vaporization}
\usage{
hum_evap_heat(...)

\method{hum_evap_heat}{default}(temp, ...)

\method{hum_evap_heat}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{temp}{Air temperature in °C.}

\item{weather_station}{A weather_station object.}
}
\value{
Numeric. Enthalpy of vaporization in J/kg.
}
\description{
Calculates heat of evaporation for water from air temperature.
}
\details{
The enthalpy of vaporization (\eqn{L}) is the amount of heat required to convert a unit mass of a liquid into vapor without a temperature change. It is calculated using the formula:
\deqn{L = (2.5008 - 0.002372 \times T) \times 10^6}
where \eqn{T} is the temperature in °C.
}
\examples{
# Calculate enthalpy of vaporization
hum_evap_heat(temp = 25)
}
\references{
Bendix 2004, p. 261.
}
