% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{hum_precipitable_water}
\alias{hum_precipitable_water}
\alias{hum_precipitable_water.default}
\alias{hum_precipitable_water.weather_station}
\title{Precipitable water}
\usage{
hum_precipitable_water(...)

\method{hum_precipitable_water}{default}(datetime, lat, elev, temp, ...)

\method{hum_precipitable_water}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link{pres_p.default}}
  \describe{
    \item{\code{g}}{Gravitational acceleration in \eqn{m \cdot s^{-2}}, default 9.81.}
    \item{\code{rl}}{Specific gas constant for air in \eqn{m^2 \cdot s^{-2} \cdot K^{-1}}, default 287.05.}
  }}

\item{datetime}{Datetime of class \code{POSIXlt}. See \code{\link[base:as.POSIXlt]{base::as.POSIXlt()}}.
Make sure to provide the correct timezone information!}

\item{lat}{Latitude in degree.}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Air temperature in degree Celcius.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Numeric. Precipitable water in cm·grams.
}
\description{
Selects reference temperature and pressure based on location and season.
Then calculates precipitable water.
}
\details{
Latitude <= 30 degrees is defined as tropic; <= 60 is temperate; others is subarctic.
Summer is defined as April to September in the northern hemisphere.

Precipitable water (\eqn{PW}) is the total amount of water vapor in a column of air from the surface to the top of the atmosphere. It is calculated using reference temperature and pressure values based on location and season.
}
\examples{
# Calculate precipitable water
hum_precipitable_water(datetime = as.POSIXlt("2022-07-15"), lat = 50, elev = 100, temp = 20)
}
\references{
Bendix 2004, p. 246.
}
