% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{hum_specific}
\alias{hum_specific}
\alias{hum_specific.default}
\alias{hum_specific.weather_station}
\title{Specific humidity}
\usage{
hum_specific(...)

\method{hum_specific}{default}(rh, temp, elev, ...)

\method{hum_specific}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{rh}{Relative humidity in \%.}

\item{temp}{Temperature in °C.}

\item{elev}{Elevation above sea level in m.}

\item{weather_station}{A weather_station object.}
}
\value{
Numeric. Specific humidity in kg/kg.
}
\description{
Calculates specific humidity from vapor pressure and air pressure.
}
\details{
Specific humidity (\eqn{q}) is the ratio of the mass of water vapor to the total mass of the air parcel. It is calculated from the vapor pressure and air pressure using the formula:
\deqn{q = 0.622 \times \frac{pvapor}{p}}
where \eqn{pvapor} is the vapor pressure and \eqn{p} is the air pressure.
}
\examples{
# Calculate specific humidity
hum_specific(rh = 70, temp = 25, elev = 100)
}
\references{
Bendix 2004, p. 262.
}
