% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressure.R
\name{pres_air_density}
\alias{pres_air_density}
\alias{pres_air_density.weather_station}
\alias{pres_air_density.default}
\title{Air Density}
\usage{
pres_air_density(...)

\method{pres_air_density}{weather_station}(weather_station, ...)

\method{pres_air_density}{default}(elev, temp, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{weather_station}{A weather_station object.}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Air temperature in degrees C.}
}
\value{
Air density in kg/m³.
}
\description{
Calculates the air density.
}
\details{
The air density (\eqn{\rho}) is calculated using the formula:
\deqn{\rho = \frac{p \cdot 100}{R \cdot T}}
where:
\eqn{p} is the air pressure in hPa,
\eqn{R} is the specific gas constant for air (287.05 m²/s²/K), and
\eqn{T} is the temperature in Kelvin (K).
}
\examples{
# Calculate air density at an elevation of 500 meters and temperature of 15°C
pres_air_density(elev = 500, temp = 15)
}
