% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_diffuse_in}
\alias{rad_diffuse_in}
\alias{rad_diffuse_in.default}
\alias{rad_diffuse_in.weather_station}
\title{Incoming Diffused Radiation}
\usage{
rad_diffuse_in(...)

\method{rad_diffuse_in}{default}(datetime, lon, lat, elev, temp, slope, exposition, valley, ...)

\method{rad_diffuse_in}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime of class \code{POSIXlt}. See \code{\link[base:as.POSIXlt]{base::as.POSIXlt()}}.
Make sure to provide the correct timezone information!}

\item{lon}{Longitude in degree.}

\item{lat}{Latitude in degree.}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Air temperature in degree Celcius.}

\item{slope}{Slope in degree.}

\item{exposition}{Exposition in degree.}

\item{valley}{Is the position in a valley (\code{TRUE}) or on a slope (\code{FALSE})?}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Diffused shortwave incoming radiation in W/m².
}
\description{
Calculates the diffused shortwave incoming radiation.
}
\details{
The diffused shortwave incoming radiation (\eqn{D_{in}}) is calculated using the formula:
\deqn{D_{in} = 0.5 \cdot [(1 - (1 - \text{vapor}) - (1 - \text{ozone})) \cdot SW_{toa} - SW_{in}] \cdot \text{sky\_view} \cdot (1 + \cos(\theta)^2 \cdot \sin(\phi)^3)}
where:
\eqn{\text{vapor}} is the vapor transmission,
\eqn{\text{ozone}} is the ozone transmission,
\eqn{SW_{toa}} is the shortwave radiation at the top of the atmosphere,
\eqn{SW_{in}} is the shortwave incoming radiation,
\eqn{\text{sky\_view}} is the sky view factor,
\eqn{\theta} is the terrain angle, and
\eqn{\phi} is the solar angle.
}
\examples{
# Calculate diffused shortwave incoming radiation
rad_diffuse_in(datetime = Sys.time(), lon = 10, lat = 50, elev = 100, temp = 15,
               slope = 5, exposition = 180, valley = FALSE)
}
\references{
Bendix 2004, p. 58 eq. 3.14, p. 55 eq. 3.9.
}
