% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_diffuse_out}
\alias{rad_diffuse_out}
\alias{rad_diffuse_out.default}
\alias{rad_diffuse_out.weather_station}
\title{Diffused Outgoing Radiation}
\usage{
rad_diffuse_out(...)

\method{rad_diffuse_out}{default}(
  datetime,
  lon,
  lat,
  elev,
  temp,
  slope,
  exposition,
  valley,
  surface_type,
  ...
)

\method{rad_diffuse_out}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime of class \code{POSIXlt}. See \code{\link[base:as.POSIXlt]{base::as.POSIXlt()}}.
Make sure to provide the correct timezone information!}

\item{lon}{Longitude in degree.}

\item{lat}{Latitude in degree.}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Air temperature in degree Celcius.}

\item{slope}{Slope in degree.}

\item{exposition}{Exposition in degree.}

\item{valley}{Is the position in a valley (\code{TRUE}) or on a slope (\code{FALSE})?}

\item{surface_type}{Surface type.
Allowed values are: field, acre, lawn, street, agriculture, settlement, coniferous forest, deciduous forest, mixed forest, city, water, shrub.
EXCEPTION: for functions related to Priestley-Taylor methods,
allowed values are: field, bare soil, coniferous forest, water, wetland, spruce forest.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Reflected diffused incoming radiation in W/m².
}
\description{
Calculates the reflected diffused incoming radiation.
}
\details{
The reflected diffused incoming radiation (\eqn{D_{out}}) is calculated using the formula:
\deqn{D_{out} = D_{in} \cdot \alpha}
where:
\eqn{D_{in}} is the diffused incoming radiation,
\eqn{\alpha} is the albedo of the surface.
}
\examples{
# Calculate reflected diffused incoming radiation
rad_diffuse_out(datetime = Sys.time(), lon = 10, lat = 50, elev = 100, temp = 15,
                slope = 5, exposition = 180, valley = FALSE, surface_type = "lawn")
}
\references{
Bendix 2004, p. 45 eq. 3.1.
}
