% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_lw_bal}
\alias{rad_lw_bal}
\alias{rad_lw_bal.default}
\alias{rad_lw_bal.weather_station}
\title{Longwave Radiation Balance}
\usage{
rad_lw_bal(...)

\method{rad_lw_bal}{default}(temp, rh, slope, valley, surface_type, surface_temp, ...)

\method{rad_lw_bal}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{temp}{Air temperature in degree Celcius.}

\item{rh}{Relative humidity in \%.}

\item{slope}{Slope in degree.}

\item{valley}{Is the position in a valley (\code{TRUE}) or on a slope (\code{FALSE})?}

\item{surface_type}{Surface type.
Allowed values are: field, acre, lawn, street, agriculture, settlement, coniferous forest, deciduous forest, mixed forest, city, water, shrub.
EXCEPTION: for functions related to Priestley-Taylor methods,
allowed values are: field, bare soil, coniferous forest, water, wetland, spruce forest.}

\item{surface_temp}{Surface temperature in degree Celcius.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Longwave radiation balance in W/m².
}
\description{
Calculates the sum of longwave incoming and outgoing radiation.
}
\details{
The longwave radiation balance (\eqn{R_{lw}}) is calculated as:
\deqn{R_{lw} = LW_{in} - LW_{out}}
where:
\eqn{LW_{in}} is the longwave incoming radiation,
\eqn{LW_{out}} is the longwave outgoing radiation.
}
\examples{
# Calculate longwave radiation balance
rad_lw_bal(temp = 15, rh = 60, slope = 5, valley = FALSE, surface_type = "lawn", surface_temp = 15)
}
\references{
Bendix 2004, p. 68. eq. 3.25.
}
