% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_lw_out}
\alias{rad_lw_out}
\alias{rad_lw_out.default}
\alias{rad_lw_out.weather_station}
\title{Longwave Outgoing Radiation}
\usage{
rad_lw_out(...)

\method{rad_lw_out}{default}(surface_type, surface_temp, ..., sigma = sigma_default)

\method{rad_lw_out}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{surface_type}{Surface type.
Allowed values are: field, acre, lawn, street, agriculture, settlement, coniferous forest, deciduous forest, mixed forest, city, water, shrub.
EXCEPTION: for functions related to Priestley-Taylor methods,
allowed values are: field, bare soil, coniferous forest, water, wetland, spruce forest.}

\item{surface_temp}{Surface temperature in degree Celcius.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Longwave outgoing radiation in W/m².
}
\description{
Calculates the longwave radiation of the surface.
}
\details{
The longwave outgoing radiation (\eqn{LW_{out}}) is calculated as:
\deqn{LW_{out} = \epsilon \cdot \sigma \cdot T_{surface}^4}
where:
\eqn{\epsilon} is the emissivity of the surface,
\eqn{\sigma} is the Stefan-Boltzmann constant, and
\eqn{T_{surface}} is the surface temperature in Kelvin.
}
\examples{
# Calculate longwave outgoing radiation
rad_lw_out(surface_type = "lawn", surface_temp = 15)
}
\references{
Bendix 2004, p. 66 eq. 3.20.
}
