% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_sw_in}
\alias{rad_sw_in}
\alias{rad_sw_in.default}
\alias{rad_sw_in.weather_station}
\title{Shortwave Incoming Radiation}
\usage{
rad_sw_in(...)

\method{rad_sw_in}{default}(datetime, lon, lat, elev, temp, slope, exposition, ...)

\method{rad_sw_in}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime of class \code{POSIXlt}. See \code{\link[base:as.POSIXlt]{base::as.POSIXlt()}}.
Make sure to provide the correct timezone information!}

\item{lon}{Longitude in degree.}

\item{lat}{Latitude in degree.}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Air temperature in degree Celcius.}

\item{slope}{Slope in degree.}

\item{exposition}{Exposition in degree.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Shortwave incoming radiation in W/m².
}
\description{
Calculates the direct shortwave incoming radiation.
}
\details{
The shortwave incoming radiation (\eqn{SW_{in}}) is calculated using the formula:
\deqn{SW_{in} = SW_{toa} \cdot 0.9751 \cdot T_{total} / \sin(E) \cdot \cos(\theta)}
where:
\eqn{SW_{toa}} is the shortwave radiation at the top of the atmosphere,
\eqn{T_{total}} is the total atmospheric transmission,
\eqn{E} is the solar elevation angle,
\eqn{\theta} is the terrain angle.
}
\examples{
# Calculate shortwave incoming radiation
rad_sw_in(datetime = Sys.time(), lon = 10, lat = 50, elev = 100, temp = 15,
          slope = 5, exposition = 180)
}
\references{
Bendix 2004, p. 46 eq. 3.3, p. 52 eq. 3.8.
}
