% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_sw_toa}
\alias{rad_sw_toa}
\alias{rad_sw_toa.default}
\alias{rad_sw_toa.weather_station}
\title{Shortwave Radiation at Top of Atmosphere}
\usage{
rad_sw_toa(...)

\method{rad_sw_toa}{default}(datetime, lon, lat, ..., sol_const = 1361)

\method{rad_sw_toa}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime of class \code{POSIXlt}. See \code{\link[base:as.POSIXlt]{base::as.POSIXlt()}}.
Make sure to provide the correct timezone information!}

\item{lon}{Longitude in degree.}

\item{lat}{Latitude in degree.}

\item{sol_const}{Solar radiation constant in W/m², default is 1361.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Shortwave radiation at top of atmosphere in W/m².
}
\description{
Calculates the shortwave radiation at the top of the atmosphere without the influence of the atmosphere.
}
\details{
The shortwave radiation at the top of the atmosphere (\eqn{SW_{toa}}) is calculated using the formula:
\deqn{SW_{toa} = S \cdot E \cdot \sin(E)}
where:
\eqn{S} is the solar constant (default 1361 W/m²),
\eqn{E} is the eccentricity correction factor,
\eqn{E} is the solar elevation angle.
}
\examples{
# Calculate shortwave radiation at top of atmosphere
rad_sw_toa(datetime = Sys.time(), lon = 10, lat = 50)
}
\references{
Bendix 2004, p. 244.
}
