% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensible.R
\name{sensible_bowen}
\alias{sensible_bowen}
\alias{sensible_bowen.default}
\alias{sensible_bowen.weather_station}
\title{Sensible Heat using Bowen Method}
\usage{
sensible_bowen(...)

\method{sensible_bowen}{default}(
  t1,
  t2,
  hum1,
  hum2,
  z1 = 2,
  z2 = 10,
  elev,
  rad_bal,
  soil_flux,
  cap = NULL,
  ...
)

\method{sensible_bowen}{weather_station}(weather_station, cap = NULL, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{t1}{Temperature at lower height in °C.}

\item{t2}{Temperature at upper height in °C.}

\item{hum1}{Relative humidity at lower height in \%.}

\item{hum2}{Relative humidity at upper height in \%.}

\item{z1}{Lower height of measurement in m.}

\item{z2}{Upper height of measurement in m.}

\item{elev}{Elevation above sea level in m.}

\item{rad_bal}{Radiation balance in W/m².}

\item{soil_flux}{Soil flux in W/m².}

\item{cap}{The cap value to prevent division by zero. Default is NULL.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Sensible heat flux in W/m².
}
\description{
Calculates the sensible heat flux using the Bowen Method. Positive
flux signifies flux away from the surface, negative values signify flux
towards the surface.
}
\details{
The sensible heat flux (\eqn{Q_h}) using the Bowen method is calculated as:
\deqn{Q_h = \frac{(R_n - G) \cdot B}{1 + B}}
where:
\eqn{R_n} is the net radiation,
\eqn{G} is the soil heat flux, and
\eqn{B} is the Bowen ratio.

The Bowen ratio (\eqn{B}) is calculated as:
\deqn{B = \frac{\gamma}{L_v} \cdot \frac{\Delta T}{\Delta q}}
where:
\eqn{\gamma} is the psychrometric constant,
\eqn{L_v} is the latent heat of vaporization,
\eqn{\Delta T} is the temperature gradient, and
\eqn{\Delta q} is the moisture gradient.

When \eqn{1 + B} results in values close to zero, the sensible heat flux can become unrealistically high.
To prevent this, a cap parameter can be set.
The cap parameter ensures that \eqn{1 + B} does not get too close to zero by setting a minimum allowable value.
}
\examples{
# Calculate sensible heat flux using the Bowen method
sensible_bowen(t1 = 20, t2 = 15, hum1 = 80, hum2 = 60, z1 = 2, z2 = 10, elev = 100, rad_bal = 200, soil_flux = 50, cap = 1)
}
\references{
Bendix 2004, p. 221, eq. 9.21
}
