% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensible.R
\name{sensible_monin}
\alias{sensible_monin}
\alias{sensible_monin.default}
\alias{sensible_monin.weather_station}
\title{Sensible Heat using Monin-Obukhov length}
\usage{
sensible_monin(...)

\method{sensible_monin}{default}(
  t1,
  t2,
  z1 = 2,
  z2 = 10,
  v1,
  v2,
  elev,
  cap = NULL,
  surface_type = NULL,
  obs_height = NULL,
  ...
)

\method{sensible_monin}{weather_station}(weather_station, cap = NULL, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{t1}{Air temperature at lower height in °C.}

\item{t2}{Air temperature at upper height in °C.}

\item{z1}{Lower height of measurement in m.}

\item{z2}{Upper height of measurement in m (Use highest point of measurement as values are less disturbed).}

\item{v1}{Windspeed at lower height (e.g. height of anemometer) in m/s.}

\item{v2}{Windspeed at upper height in m/s.}

\item{elev}{Elevation above sea level in m.}

\item{cap}{The maximum absolute value for the stability parameter \eqn{s_1}. Default is NULL.}

\item{surface_type}{Type of surface. Options: r surface_properties$surface_type}

\item{obs_height}{Height of obstacle in meters (m).}

\item{weather_station}{Object of class weather_station}
}
\value{
Sensible heat flux in W/m².
}
\description{
Calculates the sensible heat flux using the Monin-Obukhov length. Positive
flux signifies flux away from the surface, negative values signify flux
towards the surface.
}
\details{
The sensible heat flux (\eqn{Q_h}) using the Monin-Obukhov method is calculated as:
\deqn{Q_h = - \rho \cdot c_p \cdot \frac{k \cdot u_* \cdot z_2}{\phi_h} \cdot \frac{\Delta \theta}{\Delta z}}
where:
\eqn{\rho} is the air density,
\eqn{c_p} is the specific heat capacity of air,
\eqn{k} is the von Kármán constant,
\eqn{u_*} is the friction velocity,
\eqn{\phi_h} is the stability correction function for heat,
\eqn{\Delta \theta} is the potential temperature gradient, and
\eqn{\Delta z} is the height difference between measurements.

The stability correction function for heat (\eqn{\phi_h}) is calculated using the gradient Richardson number (\eqn{Ri_g}) and the stability parameter (\eqn{s_1}).
The stability parameter is the ratio of the higher measurement height and the Monin-Obukhov length.
With Monin-Obukhov length values close to zero, the ratio can result in very high values, which is why the stability parameter (\eqn{s_1}) is capped.
The default cap is set to NULL.
}
\examples{
# Calculate sensible heat flux using the Monin-Obukhov method
sensible_monin(t1 = 20, t2 = 15, z1 = 2, z2 = 10, v1 = 3, v2 = 5, elev = 100, surface_type = "lawn")
}
\references{
Bendix 2004, p. 77, eq. 4.6,

Foken 2016, p. 362: Businger
}
