% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\name{soil_thermal_cond}
\alias{soil_thermal_cond}
\alias{soil_thermal_cond.default}
\alias{soil_thermal_cond.weather_station}
\title{Soil Thermal Conductivity}
\usage{
soil_thermal_cond(...)

\method{soil_thermal_cond}{default}(texture, moisture, ...)

\method{soil_thermal_cond}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{texture}{Soil texture. Either "sand", "peat" or "clay".}

\item{moisture}{Soil moisture content in cubic meters per cubic meter.}

\item{weather_station}{A weather_station object.}
}
\value{
Soil thermal conductivity in W/m/K.
}
\description{
Calculates soil thermal conductivity from soil texture and soil moisture.
Works by linearly interpolating thermal conductivity based on measured data.
Moisture is supplied as m3 m-3 and converted internally to volume percent.
}
\details{
The thermal conductivity (\eqn{\lambda}) of the soil is determined based on its texture and moisture content.
The values are interpolated from measured data for different soil types.
Values outside the tabulated moisture domain return \code{NA}.
}
\examples{
# Calculate soil thermal conductivity
soil_thermal_cond(texture = "sand", moisture = 0.25)
}
\references{
Bendix 2004, p. 254.
}
