% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_declination}
\alias{sol_declination}
\alias{sol_declination.default}
\alias{sol_declination.weather_station}
\title{Solar declination}
\usage{
sol_declination(...)

\method{sol_declination}{default}(datetime, ...)

\method{sol_declination}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Named station fields, site parameters or model assumptions.}

\item{datetime}{POSIXlt or POSIXct date-time vector.}

\item{weather_station}{A weather_station object.}
}
\value{
Degree.
}
\description{
Solar declination
}
\details{
The solar declination (\eqn{\delta}) is the angle between the rays of the sun and the plane of the Earth's equator. It is calculated as:
\deqn{\delta = \arcsin(\sin(23.44^\circ) \cdot \sin(L))}
where:
\eqn{L} is the ecliptic longitude.
}
\examples{
# Calculate solar declination
sol_declination(as.POSIXlt("2022-06-21"))
}
\references{
Bendix 2004, p. 243.
}
