% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_eccentricity}
\alias{sol_eccentricity}
\alias{sol_eccentricity.default}
\alias{sol_eccentricity.weather_station}
\title{Eccentricity factor}
\usage{
sol_eccentricity(...)

\method{sol_eccentricity}{default}(datetime, ...)

\method{sol_eccentricity}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime of class \code{POSIXlt}. See \code{\link[base:as.POSIXlt]{base::as.POSIXlt()}}.
Make sure to provide the correct timezone information!}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Unitless.
}
\description{
The track of Earth around the Sun is not a circle, but more like an ellipse.
}
\details{
The eccentricity factor (\eqn{E}) accounts for the elliptical shape of Earth's orbit around the Sun. It is calculated as:
\deqn{E = 1.00011 + 0.034221 \cdot \cos(D) + 0.00128 \cdot \sin(D) + 0.000719 \cdot \cos(2D) + 0.000719 \cdot \sin(2D)}
where:
\eqn{D} is the day angle in radians.
}
\examples{
# Calculate eccentricity factor
sol_eccentricity(as.POSIXlt("2022-06-21"))
}
\references{
Bendix 2004, p. 243.
}
