% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_elevation}
\alias{sol_elevation}
\alias{sol_elevation.default}
\alias{sol_elevation.weather_station}
\title{Solar elevation angle}
\usage{
sol_elevation(...)

\method{sol_elevation}{default}(datetime, lon, lat, ...)

\method{sol_elevation}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Named station fields, site parameters or model assumptions.}

\item{datetime}{POSIXlt or POSIXct date-time vector.}

\item{lon}{Longitude in degrees.}

\item{lat}{Latitude in degrees.}

\item{weather_station}{A weather_station object.}
}
\value{
Degree.
}
\description{
Solar elevation angle
}
\details{
The solar elevation angle (\eqn{h}) is the apparent angle of the sun above the horizon. It is calculated as:
\deqn{h = \arcsin(\sin(\phi) \cdot \sin(\delta) + \cos(\phi) \cdot \cos(\delta) \cdot \cos(H))}
where:
\eqn{\phi} is the latitude,
\eqn{\delta} is the solar declination, and
\eqn{H} is the hour angle.
}
\examples{
# Calculate solar elevation angle
sol_elevation(as.POSIXlt("2022-06-21"), lon = 10, lat = 50)
}
\references{
Bendix 2004, p. 243.
}
