% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_hour_angle}
\alias{sol_hour_angle}
\alias{sol_hour_angle.default}
\alias{sol_hour_angle.weather_station}
\title{Solar hour angle}
\usage{
sol_hour_angle(...)

\method{sol_hour_angle}{default}(datetime, lon, ...)

\method{sol_hour_angle}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Named station fields, site parameters or model assumptions.}

\item{datetime}{POSIXlt or POSIXct date-time vector.}

\item{lon}{Longitude in degrees.}

\item{weather_station}{A weather_station object.}
}
\value{
Degree.
}
\description{
Calculates the solar hour angle, which is the measure of time since solar noon in degrees.
}
\details{
The solar hour angle (\eqn{H}) is calculated as:
\deqn{H = 15 \cdot (T_m + E_t - 12)}
where:
\eqn{T_m} is the solar medium suntime,
\eqn{E_t} is the solar time formula.
}
\examples{
# Calculate solar hour angle
sol_hour_angle(as.POSIXlt("2022-06-21 12:00:00"), lon = 10)
}
\references{
Bendix 2004, p. 243.
}
