% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_medium_anomaly}
\alias{sol_medium_anomaly}
\alias{sol_medium_anomaly.default}
\alias{sol_medium_anomaly.weather_station}
\title{Solar medium anomaly}
\usage{
sol_medium_anomaly(...)

\method{sol_medium_anomaly}{default}(datetime, ...)

\method{sol_medium_anomaly}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime of class \code{POSIXlt}. See \code{\link[base:as.POSIXlt]{base::as.POSIXlt()}}.
Make sure to provide the correct timezone information!}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Degree.
}
\description{
Calculates the solar medium anomaly, which is the angular distance of the Earth from its perihelion.
}
\details{
The solar medium anomaly (\eqn{M}) is calculated as:
\deqn{M = 356.6 + 0.9856 \cdot J}
where:
\eqn{J} is the Julian day.
}
\examples{
# Calculate solar medium anomaly
sol_medium_anomaly(as.POSIXlt("2022-06-21"))
}
\references{
Bendix 2004, p. 243.
}
