% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_medium_suntime}
\alias{sol_medium_suntime}
\alias{sol_medium_suntime.default}
\alias{sol_medium_suntime.weather_station}
\title{Solar medium suntime}
\usage{
sol_medium_suntime(...)

\method{sol_medium_suntime}{default}(datetime, lon, ...)

\method{sol_medium_suntime}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Named station fields, site parameters or model assumptions.}

\item{datetime}{POSIXlt or POSIXct date-time vector.}

\item{lon}{Longitude in degrees.}

\item{weather_station}{A weather_station object.}
}
\value{
Hour.
}
\description{
Calculates the solar medium suntime, which is the mean solar time adjusted for the observer's longitude.
}
\details{
The solar medium suntime (\eqn{T_m}) is calculated as:
\deqn{T_m = T_{local} + \frac{lon}{15}}
where:
\eqn{T_{local}} is the local time zone,
\eqn{lon} is the longitude of the observer.
}
\examples{
# Calculate solar medium suntime
sol_medium_suntime(as.POSIXlt("2022-06-21 12:00:00"), lon = 10)
}
\references{
Bendix 2004, p. 243.
}
