% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_aerosol}
\alias{trans_aerosol}
\alias{trans_aerosol.default}
\alias{trans_aerosol.weather_station}
\title{Transmittance due to aerosols}
\usage{
trans_aerosol(...)

\method{trans_aerosol}{default}(datetime, lon, lat, elev, temp, ..., vis = vis_default)

\method{trans_aerosol}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{POSIXlt object, date and time of the observation.}

\item{lon}{Longitude in decimal degrees.}

\item{lat}{Latitude in decimal degrees.}

\item{elev}{Elevation above sea level in meters.}

\item{temp}{Air temperature in °C.}

\item{vis}{Visibility in km, default \code{vis_default}.}
}
\value{
Transmittance ratio from 0 to 1, unitless.
}
\description{
Calculates transmittance due to fine particles in the air.
}
\details{
The transmittance due to aerosols is calculated using the formula:
\deqn{T_{aerosol} = \exp(-x^{0.873} \cdot (1 + x - x^{0.7088}) \cdot M_{abs}^{0.9108})}
where \eqn{x} is a function of the visibility and \eqn{M_{abs}} is the absolute optical air mass.
}
\examples{
# Calculate transmittance due to aerosols
trans_aerosol(datetime = as.POSIXlt("2023-08-06 12:00:00", tz = "UTC"), lon = 8.68, lat = 50.77, elev = 100, temp = 20, vis = 50)
}
\references{
Bendix 2004, p. 246.
}
