% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_air_mass_rel}
\alias{trans_air_mass_rel}
\alias{trans_air_mass_rel.default}
\alias{trans_air_mass_rel.weather_station}
\title{Relative optical air mass}
\usage{
trans_air_mass_rel(...)

\method{trans_air_mass_rel}{default}(datetime, lon, lat, ...)

\method{trans_air_mass_rel}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{POSIXlt object, date and time of the observation.}

\item{lon}{Longitude in decimal degrees.}

\item{lat}{Latitude in decimal degrees.}
}
\value{
Relative optical air mass, unitless.
}
\description{
Calculates the relative optical air mass.
}
\details{
The relative optical air mass is calculated using the formula:
\deqn{M_{rel} = \frac{1}{\sin(elevation) + 1.5 \cdot elevation^{-0.72}}}
where \eqn{elevation} is the solar elevation angle in degrees.
}
\examples{
# Calculate relative optical air mass
trans_air_mass_rel(datetime = as.POSIXlt("2023-08-06 12:00:00", tz = "UTC"), lon = 8.68, lat = 50.77)
}
\references{
Bendix 2004, p. 246.
}
