% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_rayleigh}
\alias{trans_rayleigh}
\alias{trans_rayleigh.default}
\alias{trans_rayleigh.weather_station}
\title{Transmittance due to rayleigh scattering}
\usage{
trans_rayleigh(...)

\method{trans_rayleigh}{default}(datetime, lon, lat, elev, temp, ...)

\method{trans_rayleigh}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{POSIXlt object, date and time of the observation.}

\item{lon}{Longitude in decimal degrees.}

\item{lat}{Latitude in decimal degrees.}

\item{elev}{Elevation above sea level in meters.}

\item{temp}{Air temperature in °C.}
}
\value{
Transmittance ratio from 0 to 1, unitless.
}
\description{
Calculates transmittance due to Rayleigh scattering.
}
\details{
The transmittance due to Rayleigh scattering is calculated using the formula:
\deqn{T_{rayleigh} = \exp(-0.0903 \cdot M_{abs}^{0.84} \cdot (1 + M_{abs} - M_{abs}^{1.01}))}
where \eqn{M_{abs}} is the absolute optical air mass.
}
\examples{
# Calculate transmittance due to rayleigh scattering
trans_rayleigh(datetime = as.POSIXlt("2023-08-06 12:00:00", tz = "UTC"), lon = 8.68, lat = 50.77, elev = 100, temp = 20)
}
\references{
Bendix 2004, p. 245.
}
