% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbulent_flux.R
\name{turb_flux_grad_rich_no}
\alias{turb_flux_grad_rich_no}
\alias{turb_flux_grad_rich_no.default}
\alias{turb_flux_grad_rich_no.weather_station}
\title{Gradient-Richardson-Number}
\usage{
turb_flux_grad_rich_no(...)

\method{turb_flux_grad_rich_no}{default}(t1, t2, z1 = 2, z2 = 10, v1, v2, elev, ...)

\method{turb_flux_grad_rich_no}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{t1}{Temperature at lower height (e.g. height of anemometer) in °C.}

\item{t2}{Temperature at upper height in degrees C.}

\item{z1}{Lower height of measurement (e.g. height of anemometer) in m.}

\item{z2}{Upper height of measurement in m.}

\item{v1}{Windspeed at lower height (e.g. height of anemometer) in m/s.}

\item{v2}{Windspeed at upper height in m/s.}

\item{elev}{Elevation above sea level in m.}

\item{weather_station}{Object of class weather_station}
}
\value{
Gradient-Richardson-Number.
}
\description{
Calculation of the Gradient-Richardson-Number. The number represents the
stability of the atmosphere. Negative values signify unstable conditions,
positive values signify stable conditions, whereas values around zero represent
neutral conditions.
}
\references{
Bendix 2004, p. 43, eq. 2.5
}
