% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbulent_flux.R
\name{turb_flux_stability}
\alias{turb_flux_stability}
\alias{turb_flux_stability.default}
\alias{turb_flux_stability.weather_station}
\title{Stability}
\usage{
turb_flux_stability(...)

\method{turb_flux_stability}{default}(grad_rich_no, ...)

\method{turb_flux_stability}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{grad_rich_no}{Gradient-Richardson-Number}

\item{weather_station}{Object of class weather_station}
}
\value{
Gradient-Richardson-Number.
}
\description{
Conversion of Gradient-Richardson-Number to stability string.
}
\references{
Based on Bendix 2004, p.43, picture 2.10
}
