% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbulence.R
\name{turb_roughness_length}
\alias{turb_roughness_length}
\alias{turb_roughness_length.default}
\alias{turb_roughness_length.weather_station}
\title{Roughness length}
\usage{
turb_roughness_length(...)

\method{turb_roughness_length}{default}(surface_type = NULL, obs_height = NULL, ...)

\method{turb_roughness_length}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{surface_type}{Type of surface. Options: r surface_properties$surface_type}

\item{obs_height}{Height of obstacle in meters (m).}

\item{weather_station}{Object of class weather_station}
}
\value{
Numeric. Roughness length in meters (m).

Numeric. Roughness length in meters (m).
}
\description{
Calculates the roughness length of a surface based on the obstacle height or the type of the surface.
}
\details{
Possible surface types are:
"field", "acre", "lawn", "street", "agriculture", "settlement", "coniferous forest", "deciduous forest", "mixed forest", "city", "water", "shrub".

You need to specify only one of \code{surface_type} or \code{obs_height}.

This function calculates the roughness length (\eqn{z_0}) of a surface. The roughness length is a measure of the roughness of the surface,
which affects the wind profile near the ground. It can be calculated either based on the height of obstacles on the surface or by specifying
the type of surface.

When the obstacle height (\code{obs_height}) is provided, the roughness length is calculated as 10\% of the obstacle height.

When the surface type (\code{surface_type}) is provided, the roughness length is looked up from predefined values.
}
\examples{
# Calculates roughness length based on obstacle height
turb_roughness_length(obs_height = 10)

# Calculate roughness length based on surface type
turb_roughness_length(surface_type = "deciduous forest")

}
\references{
Bendix 2004, p. 239
}
